+++ microbench_1N/launch_ipc_msg_queue.sh
@@ -10,6 +10,8 @@
 
 
 MEMORY_DIR="memory_conso"
+NB_THREADS_PER_CORE=2
+
 
 # get arguments
 if [ $# -eq 6 ]; then
@@ -42,6 +44,13 @@ else
    exit 0
 fi
 
+OUTPUT_DIR="microbench_ipc_msg_queue_${NB_CONSUMERS}consumers_${DURATION_XP}sec_${MSG_SIZE}B_${NB_QUEUES}queues_msg_max_size_${MESSAGE_MAX_SIZE}B"
+
+if [ -d $OUTPUT_DIR ]; then
+   echo IPC msg queue ${NB_CONSUMERS} consumers, ${DURATION_XP} sec, ${MSG_SIZE}B ${NB_QUEUES} queues ${MESSAGE_MAX_SIZE}B for msg max size already done
+   exit 0
+fi
+
 
 rm -rf $MEMORY_DIR && mkdir $MEMORY_DIR
 
@@ -61,12 +70,27 @@ if [ -z $NO_COMPILE ]; then
 fi
 
 ./get_memory_usage.sh  $MEMORY_DIR &
-./bin/ipc_msg_queue_microbench -r $NB_CONSUMERS -s $MSG_SIZE -t $DURATION_XP
+./bin/ipc_msg_queue_microbench -r $NB_CONSUMERS -s $MSG_SIZE -t $DURATION_XP &
+
+sleep 5
+
+sudo ./profiler/profiler-sampling &
+
+sleep $DURATION_XP
+sudo pkill profiler
 
 ./stop_all.sh
 
 # save files
-OUTPUT_DIR="microbench_ipc_msg_queue_${NB_CONSUMERS}consumers_${DURATION_XP}sec_${MSG_SIZE}B_${NB_QUEUES}queues_msg_max_size_${MESSAGE_MAX_SIZE}B"
 mkdir $OUTPUT_DIR
 mv $MEMORY_DIR $OUTPUT_DIR/
 mv statistics*.log $OUTPUT_DIR/
+
+sudo chown bft:bft /tmp/perf.data.*
+for c in $(seq 0 ${NB_CONSUMERS}); do
+   cid=$(( $c * $NB_THREADS_PER_CORE ))
+   for e in 0 1 2; do
+      ./profiler/parser-sampling /tmp/perf.data.* --c ${cid} --base-event ${e} --app ipc_msg_queue_m > $OUTPUT_DIR/perf_core_${cid}_event_${e}.log
+   done
+done
+rm /tmp/perf.data.* -f
